module procedure references and definition switching.
Procedure declarations in (sub)module interfaces and their corresponding definitions in submodules are
now linked together. This links is followed when collecting references, so that all references to both
are returned when querying for either.

Defintiion lookup has also been modified so that it will toggle between the interface declaration and
definition when the cursor lookup is already on one or the other.